add auth - group

+++ schedule-admin/src/main/java/com/wjs/schedule/service/job/impl/CuckooGroupServiceImpl.java
@@ -2,20 +2,23 @@ package com.wjs.schedule.service.job.impl;
 
 import java.util.List;
 
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.wjs.schedule.dao.auth.CuckooAuthJobgrpMapper;
 import com.wjs.schedule.dao.exec.CuckooJobDetailMapper;
 import com.wjs.schedule.dao.exec.CuckooJobGroupMapper;
 import com.wjs.schedule.domain.exec.CuckooJobDetailCriteria;
 import com.wjs.schedule.domain.exec.CuckooJobGroup;
 import com.wjs.schedule.domain.exec.CuckooJobGroupCriteria;
+import com.wjs.schedule.enums.CuckooUserAuthType;
 import com.wjs.schedule.exception.BaseException;
+import com.wjs.schedule.service.auth.CuckooAuthService;
 import com.wjs.schedule.service.job.CuckooGroupService;
-import com.wjs.schedule.vo.job.JobGroup;
-import com.wjs.util.bean.PropertyUtil;
+import com.wjs.schedule.vo.auth.CuckooLogonInfo;
 
 @Service("cuckooGroupService")
 public class CuckooGroupServiceImpl implements CuckooGroupService {
@@ -26,24 +29,43 @@ public class CuckooGroupServiceImpl implements CuckooGroupService {
 	@Autowired
 	CuckooJobDetailMapper cuckooJobDetailMapper;
 	
+	@Autowired
+	CuckooAuthService cuckooAuthService;
+	
+	
 	@Override
 	@Transactional
-	public Long addGroup(JobGroup group) {
+	public Long addGroup(CuckooJobGroup cuckooJobGroup) {
 		
-		if(null == group || StringUtils.isEmpty(group.getGroupName())){
+		if(null == cuckooJobGroup || StringUtils.isEmpty(cuckooJobGroup.getGroupName())){
 			throw new BaseException("group name should not be null");
 		}
-		group.setId(null);
-		CuckooJobGroup cuckGroup = new CuckooJobGroup();
-		PropertyUtil.copyProperties(cuckGroup, group);
-		cuckooJobGroupMapper.insertSelective(cuckGroup);
-		return cuckooJobGroupMapper.lastInsertId();
+		
+		if(CuckooUserAuthType.GUEST.getValue().equals(cuckooAuthService.getLogonInfo().getCuckooUserAuthType().getValue())){
+			throw new BaseException("Guest Users had no right to add group");
+		}
+		
+		cuckooJobGroup.setId(null);
+		cuckooJobGroupMapper.insertSelective(cuckooJobGroup);
+		Long groupId = cuckooJobGroupMapper.lastInsertId();
+		
+		// 谁新增的，有直接所有权限
+		cuckooAuthService.addAuthJobgrp(cuckooAuthService.getLogonInfo().getId(), groupId);
+		cuckooAuthService.refreshAuth(cuckooAuthService.getLogonInfo());
+		return groupId;
 	}
 
 	@Override
-	public List<CuckooJobGroup> selectAllGroup() {
-		
-		return cuckooJobGroupMapper.selectByExample(new CuckooJobGroupCriteria());
+	public List<CuckooJobGroup> listAllGroup() {
+
+		CuckooLogonInfo loginInfo = cuckooAuthService.getLogonInfo();
+		CuckooJobGroupCriteria crt = new CuckooJobGroupCriteria();
+
+		if (CollectionUtils.isNotEmpty(loginInfo.getReadableGroupIds())) {
+			crt.createCriteria().andIdIn(loginInfo.getReadableGroupIds());
+		}
+
+		return cuckooJobGroupMapper.selectByExample(crt);
 	}
 
 	@Override
@@ -55,11 +77,27 @@ public class CuckooGroupServiceImpl implements CuckooGroupService {
 	@Override
 	public void deleteById(Long id) {
 
+		if (!cuckooAuthService.getLogonInfo().getWritableGroupIds().contains(id)) {
+			throw new BaseException("not writable right in this group");
+		}
+
 		cuckooJobGroupMapper.deleteByPrimaryKey(id);
 		CuckooJobDetailCriteria crt = new CuckooJobDetailCriteria();
 		crt.createCriteria().andGroupIdEqualTo(id);
-		
+
 		cuckooJobDetailMapper.deleteByExample(crt);
+		
+		cuckooAuthService.refreshAuth(cuckooAuthService.getLogonInfo());
+	}
+
+	@Override
+	public void updateByPk(CuckooJobGroup cuckooJobGroup) {
+
+		if (!cuckooAuthService.getLogonInfo().getWritableGroupIds().contains(cuckooJobGroup.getId())) {
+			throw new BaseException("not writable right in this group {}", cuckooJobGroup.getGroupName());
+		}
+		cuckooAuthService.refreshAuth(cuckooAuthService.getLogonInfo());
+		cuckooJobGroupMapper.updateByPrimaryKeySelective(cuckooJobGroup);
 	}
 
 }
